# openapi.yaml
openapi: 3.1.0
info:
  title: Provenonce Beats API
  version: 1.2.0
  description: Stateless public utility for anchor clock reads, proof verification, and hash timestamping.
servers:
  - url: https://beats.provenonce.dev
paths:
...
  /api/v1/beat/timestamp:
    post:
      summary: Anchor a SHA-256 hash to current beat and Solana memo
      description: utc fields in responses are Unix epoch milliseconds.
      parameters:
        - in: header
          name: X-Beats-Tier-Token
          required: false
          schema:
            type: string
          description: Optional pro-tier token for higher quotas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hash]
              properties:
                hash:
                  type: string
                  pattern: '^[0-9a-f]{64}$'
      responses:
        '200':
          description: Timestamp receipt
        '400':
          description: Invalid hash/body
        '413':
          description: Request too large
        '415':
          description: Invalid content type
        '429':
          description: Rate limited (free 5/min + 10/day, pro 30/min + 500/day per IP)
        '502':
          description: Solana write/read failure
        '503':
          description: No anchor available yet